Update fire.rpy
Dear Lord Ryan,

It is with my deepest condolences that I must inform you that there shall be more overtime added to your schedule. Alas, I have yet to commit myself to reading the "renpy guide" in full, and have therefore left you an array of notes... nay... tasks to complete. I expect the implementations to be added post-haste. You will not be compensated for the additional hours.

Signed, Aikhinneas "Ikea" Lamberton the Third.