added lee's sprites and base transforms for all characters
it's day 24.
there is no hope for any of us.
there is no hope for any of us.
there is no hope for any of us.
there is no hope for any of us.
there is no hope for any of us.
there is no hope for any of us.
there is no hope for any of us.